Moving CMC interface
