+++ rs/nervous_system/canisters/BUILD.bazel
@@ -13,6 +13,7 @@ DEPENDENCIES = [
     "//rs/rust_canisters/dfn_core",
     "//rs/types/base_types",
     "@crate_index//:candid",
+    "@crate_index//:mockall"
 ]
 
 MACRO_DEPENDENCIES = [
