+++ rs/nervous_system/common/src/lib.rs
@@ -26,7 +26,6 @@ use std::{
 };
 
 pub mod binary_search;
-pub mod cmc;
 pub mod dfn_core_stable_mem_utils;
 pub mod ledger;
 pub mod ledger_validation;
