+++ rs/nns/governance/tests/governance.rs
@@ -21,9 +21,9 @@ use fixtures::{
 use futures::future::FutureExt;
 use ic_base_types::{CanisterId, NumBytes, PrincipalId};
 use ic_crypto_sha2::Sha256;
+use ic_nervous_system_canisters::cmc::CMC;
 use ic_nervous_system_clients::canister_status::{CanisterStatusResultV2, CanisterStatusType};
 use ic_nervous_system_common::{
-    cmc::CMC,
     ledger,
     ledger::{compute_neuron_staking_subaccount_bytes, IcpLedger},
     NervousSystemError, E8, ONE_DAY_SECONDS, ONE_MONTH_SECONDS, ONE_YEAR_SECONDS,
