+++ rs/sns/governance/BUILD.bazel
@@ -20,6 +20,7 @@ DEPENDENCIES = [
     "//rs/ledger_suite/common/ledger_core",
     "//rs/ledger_suite/icp:icp_ledger",
     "//rs/ledger_suite/icrc1/ledger",
+    "//rs/nervous_system/canisters",
     "//rs/nervous_system/clients",
     "//rs/nervous_system/collections/union_multi_map",
     "//rs/nervous_system/common",
@@ -155,7 +156,6 @@ rust_canister(
     deps = DEPENDENCIES + [
         ":build_script",
         ":governance",
-        "//rs/nervous_system/canisters",
     ],
 )
 
@@ -170,7 +170,6 @@ rust_canister(
     deps = DEPENDENCIES + [
         ":build_script",
         ":governance",
-        "//rs/nervous_system/canisters",
     ],
 )
 
@@ -200,7 +199,6 @@ rust_test(
     deps = DEPENDENCIES + DEV_DEPENDENCIES + [
         ":build_script",
         ":governance",
-        "//rs/nervous_system/canisters",
     ],
 )
 
@@ -214,7 +212,6 @@ rust_test(
     deps = DEPENDENCIES + DEV_DEPENDENCIES + [
         ":build_script",
         ":governance",
-        "//rs/nervous_system/canisters",
     ],
 )
