+++ rs/sns/governance/src/governance.rs
@@ -87,8 +87,8 @@ use ic_management_canister_types_private::{
 };
 use ic_nervous_system_clients::ledger_client::ICRC1Ledger;
 use ic_nervous_system_collections_union_multi_map::UnionMultiMap;
+use ic_nervous_system_canisters::cmc::CMC;
 use ic_nervous_system_common::{
-    cmc::CMC,
     i2d,
     ledger::{self, compute_distribution_subaccount_bytes},
     NervousSystemError, ONE_DAY_SECONDS,
