+++ rs/sns/integration_tests/BUILD.bazel
@@ -14,6 +14,7 @@ BASE_DEPENDENCIES = [
     "//rs/ledger_suite/icp/ledger",
     "//rs/ledger_suite/icrc1",
     "//rs/ledger_suite/icrc1/ledger",
+    "//rs/nervous_system/canisters",
     "//rs/nervous_system/common",
     "//rs/nervous_system/common/test_keys",
     "//rs/nns/cmc",
